Separate app configuration from the view model
